COC-132-refactor(#26): timeUtils 파일명 수정
